requirements.txt
Jupyter
pymupdf 
pillow
numpy
sentence-transformers
transformers
torch
torchvision
faiss-cpu
pypdf
unstructured
openai
langchain
langchain-core
langchain-community
langchain-openai
langchain-cerebras
langchain-huggingface
pdf2image
pytesseract